refactor: move wallet accounts / balance into ledger module
