+++ src/ledger/templates/spend_detected.rs
@@ -8,7 +8,9 @@ use std::collections::HashMap;
 
 use super::shared_meta::*;
 use crate::{
-    error::*, ledger::constants::*, primitives::*, wallet::balance::WalletLedgerAccountIds,
+    error::*,
+    ledger::{constants::*, WalletLedgerAccountIds},
+    primitives::*,
 };
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
