+++ src/wallet/balance.rs
@@ -1,31 +1,6 @@
 use rust_decimal::Decimal;
-use sqlx_ledger::balance::AccountBalance;
 
-use crate::primitives::{LedgerAccountId, Satoshis};
-
-#[derive(Debug, Clone, Copy)]
-pub struct WalletLedgerAccountIds {
-    pub onchain_incoming_id: LedgerAccountId,
-    pub onchain_at_rest_id: LedgerAccountId,
-    pub onchain_outgoing_id: LedgerAccountId,
-    pub effective_incoming_id: LedgerAccountId,
-    pub effective_at_rest_id: LedgerAccountId,
-    pub effective_outgoing_id: LedgerAccountId,
-    pub fee_id: LedgerAccountId,
-    pub dust_id: LedgerAccountId,
-}
-
-#[derive(Debug)]
-pub struct WalletLedgerAccountBalances {
-    pub onchain_incoming: Option<AccountBalance>,
-    pub onchain_at_rest: Option<AccountBalance>,
-    pub onchain_outgoing: Option<AccountBalance>,
-    pub effective_incoming: Option<AccountBalance>,
-    pub effective_at_rest: Option<AccountBalance>,
-    pub effective_outgoing: Option<AccountBalance>,
-    pub fee: Option<AccountBalance>,
-    pub dust: Option<AccountBalance>,
-}
+use crate::{ledger::WalletLedgerAccountBalances, primitives::Satoshis};
 
 #[derive(Debug)]
 pub struct WalletBalanceSummary {
